requirements.txt
pysam
pyvcf